# Conflicts:
#	changelogs/minor.md

+++ system/ee/legacy/libraries/Cache/drivers/Cache_file.php
@@ -4,7 +4,7 @@
  * ExpressionEngine (https://expressionengine.com)
  *
  * @link      https://expressionengine.com/
- * @copyright Copyright (c) 2003-2021, Packet Tide, LLC (https://www.packettide.com)
+ * @copyright Copyright (c) 2003-2022, Packet Tide, LLC (https://www.packettide.com)
  * @license   https://expressionengine.com/license Licensed under Apache License, Version 2.0
  */
 
@@ -49,6 +49,10 @@ class EE_Cache_file extends CI_Driver
         }
 
         $data = unserialize(file_get_contents($this->_cache_path . $key));
+
+        if (!is_array($data)) {
+            return false;
+        }
 
         if ($data['ttl'] > 0 && ee()->localize->now > $data['time'] + $data['ttl']) {
             unlink($this->_cache_path . $key);
